refactor: 💡 exception 패키지 생성후 해당하는 클래스 이동
✅ Closes: #106
