+++ src/main/java/com/postsquad/scoup/web/signin/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.postsquad.scoup.web.signin.service;
+package com.postsquad.scoup.web.signin.exception;
 
 // TODO: 추후에 상속 구조 올바른지 파악 필요. UserNotFound면 User도메인에 속해야 하는게 아닌지?
 public class UserNotFoundException extends SignInFailedException {
