+++ src/main/java/com/postsquad/scoup/web/signin/service/SignInService.java
@@ -2,6 +2,7 @@ package com.postsquad.scoup.web.signin.service;
 
 import com.postsquad.scoup.web.signin.controller.request.SignInRequest;
 import com.postsquad.scoup.web.signin.controller.response.SignInResponse;
+import com.postsquad.scoup.web.signin.exception.UserNotFoundException;
 import com.postsquad.scoup.web.signin.mapper.SignInResponseMapper;
 import com.postsquad.scoup.web.user.domain.User;
 import com.postsquad.scoup.web.user.repository.UserRepository;
